+++ src/Service/Admin/Audiobook/AudiobooksID3TagsReaderService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Service\Admin;
+namespace App\Service\Admin\Audiobook;
 
 class AudiobooksID3TagsReaderService implements AudiobooksID3TagsReaderServiceInterface
 {
