+++ src/Service/Admin/Notification/AdminNotificationAddServiceInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Service\Admin\Notification;
+
+interface AdminNotificationAddServiceInterface
+{
+}
