+++ src/Service/Admin/Notification/AdminNotificationPatchServiceInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Service\Admin\Notification;
+
+interface AdminNotificationPatchServiceInterface
+{
+}
